requirements.txt
fastapi
uvicorn
requests
jinja2
fastapi-mail
python-dotenv
pydantic